+++ ubo_app/services/090-assistant/reducer.py
@@ -2,17 +2,21 @@
 from __future__ import annotations
 
 from dataclasses import replace
-from typing import TYPE_CHECKING, TypeAlias
+from typing import TYPE_CHECKING
 
 from redux import CompleteReducerResult, InitializationActionError
 from redux.basic_types import InitAction
 
 from ubo_app.store.services.assistant import (
+    DEFAULT_MODELS,
+    AssistantAction,
     AssistantDownloadOllamaModelAction,
     AssistantDownloadOllamaModelEvent,
     AssistantEvent,
     AssistantProcessSpeechEvent,
+    AssistantSetIsActiveAction,
     AssistantSetSelectedEngineAction,
+    AssistantSetSelectedModelAction,
     AssistantState,
 )
 from ubo_app.store.services.speech_recognition import (
@@ -22,12 +26,10 @@ from ubo_app.store.services.speech_recognition import (
 if TYPE_CHECKING:
     from redux import ReducerResult
 
-Action: TypeAlias = InitAction | SpeechRecognitionReportSpeechAction
-
 
 def reducer(
     state: AssistantState | None,
-    action: Action,
+    action: AssistantAction,
 ) -> ReducerResult[AssistantState, None, AssistantEvent]:
     if state is None:
         if isinstance(action, InitAction):
@@ -35,8 +37,18 @@ def reducer(
 
         raise InitializationActionError(action)
 
+    if isinstance(action, AssistantSetIsActiveAction):
+        return replace(state, is_active=action.is_active)
+
     if isinstance(action, AssistantSetSelectedEngineAction):
-        return replace(state, selected_engine=action.engine_name)
+        return replace(
+            state,
+            selected_engine=action.engine_name,
+            selected_model=DEFAULT_MODELS[action.engine_name],
+        )
+
+    if isinstance(action, AssistantSetSelectedModelAction):
+        return replace(state, selected_model=action.model)
 
     if isinstance(action, AssistantDownloadOllamaModelAction):
         return CompleteReducerResult(