+++ ubo_app/store/services/assistant.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from dataclasses import field
+from enum import StrEnum
 
 from immutable import Immutable
 from redux import BaseAction, BaseEvent
@@ -11,14 +12,32 @@ from ubo_app.constants import DEFAULT_ASSISTANT_OLLAMA_MODEL
 from ubo_app.utils.persistent_store import read_from_persistent_store
 
 
+class AssistantEngineName(StrEnum):
+    """Available speech recognition engines."""
+
+    OLLAMA = 'ollama'
+
+
 class AssistantAction(BaseAction):
     """Base class for assistant actions."""
+
+
+class AssistantSetIsActiveAction(AssistantAction):
+    """Action to set the assistant active state."""
+
+    is_active: bool
 
 
 class AssistantSetSelectedEngineAction(AssistantAction):
     """Action to set the selected engine."""
 
-    engine_name: str
+    engine_name: AssistantEngineName
+
+
+class AssistantSetSelectedModelAction(AssistantAction):
+    """Action to set the selected model."""
+
+    model: str
 
 
 class AssistantDownloadOllamaModelAction(AssistantAction):
@@ -44,12 +63,30 @@ class AssistantProcessSpeechEvent(AssistantEvent):
     text: str
 
 
+DEFAULT_MODELS = {
+    AssistantEngineName.OLLAMA: DEFAULT_ASSISTANT_OLLAMA_MODEL,
+}
+
+
 class AssistantState(Immutable):
     """State for the assistant service."""
 
-    selected_engine: str = field(
+    is_active: bool = field(
+        default=read_from_persistent_store(
+            'assistant:is_active',
+            default=False,
+        ),
+    )
+
+    selected_engine: AssistantEngineName = field(
         default=read_from_persistent_store(
             'assistant:selected_engine',
-            default=f'ollama:{DEFAULT_ASSISTANT_OLLAMA_MODEL}',
+            default=AssistantEngineName.OLLAMA,
+            mapper=lambda value: AssistantEngineName(value)
+            if value in AssistantEngineName.__members__
+            else AssistantEngineName.OLLAMA,
         ),
+    )
+    selected_models: dict[AssistantEngineName, str] = field(
+        default_factory=lambda: DEFAULT_MODELS,
     )