+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - refactor(speech-recognition): move the `BackgroundRunningMixin` and `NeedsSetupMixin` abstractions to the core as other modules can use them too
 - refactor(core): add `mapper` to `read_from_persistent_store` to map the read string into the object, add support for enums in `UboStore.load_object`
 - fix(core): switch to `packaging.version` for parsing Python package versions to ensure PEP 440 compliance instead of using `semver`
+- refactor(assistant): add abstractions for assistant module extending engine abstractions in the core and reimplement the service based on these abstract base classes
 
 ## Version 1.4.0